requirements.txt
anytree==2.8.0
apparmor==3.0.7
appdirs==1.4.4
attrs==20.3.0
autobahn==22.7.1
autocommand==2.2.1
Automat==20.2.0
beautifulsoup4==4.10.0
btrfsutil==5.19
CacheControl==0.12.6
canvasapi==2.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
check50==3.3.5
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
contextlib2==21.6.0
cryptography==37.0.4
Cython==0.29.32
dbus-python==1.2.18
distlib==0.3.5
distro==1.7.0
docopt==0.6.2
docutils==0.19
escrotum==1.0.1
Flask==2.2.1
future==0.18.2
hkdf==0.0.3
html5lib==1.1
httpserver==1.1.0
humanize==3.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflect==6.0.0
itsdangerous==2.1.2
jaraco.context==4.1.1
jaraco.functools==3.5.1
jaraco.text==3.9.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.1
lensfun==0.3.3
lib50==3.0.3
LibAppArmor==3.0.7
lxml==4.9.1
magic-wormhole==0.12.0
MarkupSafe==2.1.1
more-itertools==8.13.0
msgpack==1.0.3
nspektr==0.4.0
ordered-set==4.0.2
packaging==21.3
parso==0.8.3
pep517==0.12.0
pexpect==4.8.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
progress==1.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
PyGObject==3.42.2
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
resolvelib==0.8.1
retrying==1.3.3
rich @ file:///build/python-rich/src/rich-12.5.1/dist/rich-12.5.0-py3-none-any.whl
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
spake2==0.8
submit50==3.1.0
team==1.0
tenacity==8.0.1
termcolor==1.1.0
todoist-python==8.1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
trove-classifiers==2022.8.7
Twisted==21.7.0
txaio==22.2.1
txtorcon==22.0.0
typing_extensions==4.3.0
ufw==0.36.1
ujson==5.4.0
urllib3==1.26.11
validate-pyproject==0.9
webencodings==0.5.1
Werkzeug==2.2.1
wsaccel==0.6.3
xcffib==0.11.1
zope.interface==5.4.0
